Move construction test to tests/
This way we can be sure we expose everything properly.
